packages/api/pyproject.toml
[project]
name = "api"
version = "1.0.0"
description = ""
requires-python = ">=3.11,<3.12"
dependencies = [
  "dj_database_url==2.3.0",
  "Django==5.1.4",
  "django-stubs==5.1.3",
  "gunicorn==23.0.0",
  "psycopg[c]==3.2.3",
  "python-dateutil>=2.9.0.post0",
  "requests>=2.32.3",
  "redis>=5.2.1",
  "hiredis>=3.1.0",
  "celery>=5.4.0",
  "django-celery-beat>=2.7.0",
  "djangorestframework>=3.16.0",
  "markdown>=3.8",
  "django-filter>=25.1",
  "dj-rest-auth[with-social]>=7.0.1",
  "pre-commit>=4.2.0",
  "drf-spectacular>=0.28.0",
  "djangorestframework-simplejwt>=5.5.0",
  "djangorestframework-stubs>=3.15.3",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
